+++ Proyecto/Clínica/src/persistencia/proxy/Coseguro.java
@@ -2,10 +2,10 @@ package persistencia.proxy;
 
 public interface Coseguro {
    public int getCodigoCoseguro();
-   
+
    public void setCodigoCoseguro(int codigoCoseguro);
-   
-   public float getPorcentaje();
-   
-   public void setPorcentaje(float porcentaje);
-}
+
+   public double getPorcentaje();
+
+   public void setPorcentaje(double porcentaje);
+} // fin de la clase Coseguro
